pablo peter

+++ laravel12/app/Http/Controllers/MotoristaController.php
@@ -4,19 +4,25 @@ namespace App\Http\Controllers;
 
 use App\Models\Motorista;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 
 class MotoristaController extends Controller
 {
     public function index() {
+        Gate::authorize('viewAny', Motorista::class);
+        
         $motoristas = Motorista::orderBy('id','desc')->get();
         return view('motorista.index', compact('motoristas'));
     }
 
     public function create() {
+        Gate::authorize('create', Motorista::class);
+        
         return view('motorista.create');
     }
 
     public function store(Request $request) {
+        Gate::authorize('create', Motorista::class);
 
         $request->validate([
             'placa' => 'required|unique:motoristas,placa',
@@ -44,16 +50,19 @@ class MotoristaController extends Controller
             return redirect()->route('motorista.index');
         }
 
+        Gate::authorize('update', $motorista);
+
         return view('motorista.edit', compact('motorista'));
     }
 
     public function update(Request $request, string $id) {
-
         $motorista = Motorista::find($id);
 
         if(!isset($motorista)) {
             return redirect()->route('motorista.index');
         }
+
+        Gate::authorize('update', $motorista);
 
         $request->validate([
             'placa' => 'required|unique:motoristas,placa,'.$motorista->id,
@@ -74,22 +83,22 @@ class MotoristaController extends Controller
         return redirect()->route('motorista.index');
     }
 
-    public function destroy(string $id)
-{
-    $motorista = Motorista::find($id);
+    public function destroy(string $id) {
+        $motorista = Motorista::find($id);
 
-    if (!isset($motorista)) {
+        if (!isset($motorista)) {
+            return redirect()->route('motorista.index');
+        }
+
+        Gate::authorize('delete', $motorista);
+
+        foreach ($motorista->entradas as $entrada) {
+            $entrada->encerrada = true;
+            $entrada->save();
+        }
+
+        $motorista->delete();
+
         return redirect()->route('motorista.index');
     }
-
-    foreach ($motorista->entradas as $entrada) {
-        $entrada->encerrada = true;
-        $entrada->save();
-    }
-
-    $motorista->delete();
-
-    return redirect()->route('motorista.index');
-}
-
 }
